third_party/WebKit/LayoutTests/webaudio/osc-sine-sweep-snr-expected.txt

Test Sine Oscillator with Exponential Sweep

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 81.05 dB
PASS Maximum difference below threshold of 6.78 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 17608 out of 176400
PASS successfullyParsed is true

TEST COMPLETE
